updated puppy to remove counter, puppy_repo finished, moved the repos into a database folder under lib/puppy_breeder while the puppy and purchase request were moved into entities folder in the same directory. Updated the puppy_breeder.rb to add setter and getter methods. Also added instances at the bottom and stored them in the getter.
